seance 2 terminer

+++ app/src/main/java/com/example/fasanol/outerspacemanager/models/UserConnected.java
@@ -11,7 +11,7 @@ public class UserConnected {
     private int gasModifier;
     private double minerals;
     private int mineralsModifier;
-    private int points;
+    private double points;
     private String username;
 
     public UserConnected(Editable username){
